Support for NXP i.MX RT5xxS devices
The target YAML was initially generated using target-gen using CMSIS packs,
but then slightly modified as follows:
 - Manually added the FlexSPI0 memory-mapped space as non-volatile RAM
   so that probe-rs understands it as program space. The flash algorithm
   for it is unmodified from the CMSIS packs, and is apparently written
   for the official RT595S evaluation board in particular. I don't know if
   it's general enough to work with flash chips beyond the one included
   on the evaluation board.
 - Start and end addresses padded to 8 hex characters so they are all
   consistent and easier to compare with the memory map in the datasheet.
 - Core names consistently set to "cm33". The original generated YAML was
   inconsistent, with RT595S using a mixture of "cm33" and "main" that
   caused the resulting memory map to malfunction. (RT595S has a second
   processor DSP core that is also debuggable, so "cm33" is to distinguish
   the Cortex-M33 core from the DSP core, although there's no debug
   support for the DSP core here yet -- it's an Xtensa-based core.)
 - Only one "variant" per chip type. The pack defines two each but they
   have identical memory maps.

These chips have no on-board flash memory and so instead they boot into
an immutable boot ROM which then probes various other peripherals to see
if they can provide application code. Debug is disabled while the boot
ROM is running, and so a normal reset catch is non-functional on this
chip, and there's no fixed entry point to attach a breakpoint to.

Instead, the datasheet documents an address that gets read once the boot
process is complete, and recommends using a data watchpoint to catch that
read and halt the processor.

Furthermore, if the boot ROM cannot find a valid program image in any of
the places it looks the chip enters its ISP mode, in which debug remains
disabled. In that case the only way to achieve debug control of the chip
is to interact with a custom "debug mailbox" interface that can command
the ISP code to yield control to a busy loop with debug enabled, after
which we can finally halt the core.

The custom debug sequence implementation handles all of these oddities to
still achieve a reset-and-halt as long as the chip hasn't been configured
for Debug Authentication (datasheet section 60.6), which isn't supported
here at all.