[WIP] Allow targets to constrain the ranges of RTT scanning
For targets that have very large amounts of RAM, that have some RAM regions
that don't make sense for RTT, or that have multiple aliases for the same
RAM the default behavior of scanning the entirety of all RAM regions is
sub-optimal.

For those targets we'll allow them to specify the what subset of RAM to
scan for a RTT block, with the option of specifying no ranges at all if
there's no reasonable way to narrow down the search space.