mimx5xxs: Now-working debug sequences for NXP i.MX 500 series
After a considerable amount of studying the reference manual, interspersed
with trial and error, this now seems to be sufficient for us to properly
recover if the boot ROM fails to find an image to start and so enters its
ISP mode.

Software reset can sometimes leave the external flash in a state that
the boot ROM can't handle, so we're supposed to pulse a reset line to
reset the flash. However, the means of doing that is board-specific, and
so what I've implemented here is specific to the official evaluation kit
and not generalized for other boards. It's not clear how to resolve this
since debug sequences don't have any possibility for configurable
settings as they would in the CMSIS pack representation of sequences.

The flash algorithms actually also perform this reset step before doing
their main work and so this reset pulse is a bit redundant for the
flash case, but including it also in the normal reset sequence means that
it will also get done when we _just_ reset the target, without any
intention of flashing new code.